modularized, new login and signup, take out modules, mobile frinedly comps
