+++ .gitignore
@@ -11,6 +11,9 @@ pids
 *.seed
 *.pid.lock
 
+src/public/dist
+src/client/public/dist/*
+
 # Directory for instrumented libs generated by jscoverage/JSCover
 lib-cov
