+++ package.json
@@ -51,6 +51,7 @@
     "react-modal": "^1.9.7",
     "react-router": "^4.1.2",
     "react-router-dom": "^4.1.2",
+    "react-toolbox": "^2.0.0-beta.12",
     "request": "^2.81.0",
     "request-promise": "^4.2.1",
     "sequelize": "^3.30.4",
