+++ src/client/components/authentication/Login.css
@@ -10,6 +10,13 @@
 
 .submit{
   font-family: "Roboto", sans-serif;
+  width: 100%
+}
+
+.form {
+  display: flex;
+  justify-content: center;
+  flex-direction: column;
 }
 
 input {
