+++ src/client/components/authentication/SignUp.js
@@ -5,93 +5,10 @@ import {
 } from 'react-router-dom';
 import firebase from 'firebase';
 import s from './Register.css';
+import RegisterForm from './RegisterForm.js';
 import Header from '../header/Header.js';
 import Footer from '../footer/Footer.js';
 
-class RegisterForm extends React.Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      firstName: '',
-      lastName: '',
-      email: '',
-      phoneNumber: '',
-      passwordInput: '',
-      photoURL: '',
-      facebookData: '',
-    };
-
-    this.handleChange = this.handleChange.bind(this);
-    this.handleSubmit = this.handleSubmit.bind(this);
-  }
-
-  handleChange(event) {
-    this.setState({value: event.target.value});
-  }
-
-  handleSubmit(event) {
-    alert('A name was submitted: ' + this.state.value);
-    event.preventDefault();
-  }
-
-  render() {
-    return (
-      <form onSubmit={this.handleSubmit}>
-         <input
-           value={this.state.firstName}
-           placeholder='First Name'
-           onChange={(firstName) => this.setState({firstName})}
-           autoCapitalize='none'
-           type="text"
-         />
-         <div>
-         <input
-           value={this.state.lastName}
-           placeholder='Last Name'
-           onChange={(lastName) => this.setState({lastName})}
-           autoCapitalize='none'
-           type="text"
-         />
-       </div>
-       <div>
-         <input
-           value={this.state.email}
-           placeholder='Email'
-           onChange={(email) => this.setState({email})}
-           autoCapitalize='none'
-           type="email"
-         />
-       </div>
-       <div>
-         <input
-           keyboardType='phone-pad'
-           placeholder='(XXX)XXX-XXXX'
-           onChange={(phoneNumber) => this.setState({phoneNumber})}
-           value={this.state.phoneNumber}
-           autoCapitalize='none'
-         />
-         <div>
-         <input
-          placeholder='password'
-          onChange={(passwordInput) => this.setState({passwordInput})}
-          value={this.state.passwordInput}
-          autoCapitalize='none'
-        />
-      </div>
-      <div>
-        <input
-          placeholder='repeat password'
-          autoCapitalize='none'
-        />
-      </div>
-       </div>
-       <div>
-        <input type="submit" value="Submit" className={s.submit}/>
-      </div>
-      </form>
-    );
-  }
-}
 
 class SignUp extends React.Component {
   constructor(props) {
@@ -133,7 +50,6 @@ class SignUp extends React.Component {
         <Header />
         <div className={s.root}>
           <div className={s.container}>
-            <h1>Register</h1>
             <RegisterForm />
             <button
               className={s.loginBtn}
